corrections

+++ settings.py
@@ -1,5 +1,5 @@
-import numpy
 
-SCREEN_SIZE : tuple[numpy.int16 , numpy.int16] = (1280 , 720)
-BG_COLOR : tuple[numpy.int8 , numpy.int8 , numpy.int8 ] = (255,255,255)
+
+SCREEN_SIZE : tuple[int , int] = (1280 , 720)
+BG_COLOR : tuple[int , int , int ] = (255,255,255)
 MOVEMENT_SPEED : float = 1
